auto: issue #43 Presentational split of frontend/src/components/chat/TurnActivityFeed.tsx (943 lines)
